:truck: Move fixtures into conftest of subpackage
