+++ tests/oidc/test_oidc_integration.py
@@ -1,15 +1,10 @@
-from urllib.parse import parse_qs, urlsplit
-
 from django.conf import settings
-from django.contrib.sessions.backends.db import SessionStore
 from django.http import HttpRequest
-from django.test import Client, RequestFactory
+from django.test import Client
 from django.urls import reverse
 
 import pytest
-from mozilla_django_oidc_db.config import store_config
 from mozilla_django_oidc_db.models import UserInformationClaimsSources
-from mozilla_django_oidc_db.typing import JSONObject
 from responses import RequestsMock
 
 from digid_eherkenning.oidc.models import (
@@ -25,78 +20,10 @@ from digid_eherkenning.oidc.views import (
     eh_bewindvoering_init,
     eh_init,
 )
-from tests.project.oidc_backends import MockBackend
 
 pytestmark = [
     pytest.mark.skipif(not settings.OIDC_ENABLED, reason="OIDC integration disabled"),
 ]
-
-
-@pytest.fixture(autouse=True)
-def disable_solo_cache(settings):
-    settings.SOLO_CACHE = None
-
-
-@pytest.fixture
-def auth_request(rf: RequestFactory):
-    request = rf.get("/authenticate/some-oidc", {"next": "/"})
-    session = SessionStore()
-    session.save()
-    request.session = session
-    return request
-
-
-@pytest.fixture
-def mock_auth_backend(request, mocker):
-    marker = request.node.get_closest_marker("mock_backend_claims")
-    claims: JSONObject = marker.args[0] if marker else {"sub": "some_username"}
-    mock_backend = MockBackend(claims=claims)
-    backend_path = f"{MockBackend.__module__}.{MockBackend.__qualname__}"
-    mocker.patch(
-        "django.contrib.auth._get_backends", return_value=[(mock_backend, backend_path)]
-    )
-    return mock_backend
-
-
-@pytest.fixture
-def callback(
-    request, auth_request: HttpRequest, rf: RequestFactory, client: Client, mocker
-) -> tuple[HttpRequest, Client]:
-    """
-    A django request primed by an OIDC auth request flow, ready for the callback flow.
-    """
-    mocker.patch(
-        "digid_eherkenning.oidc.views.OIDCInit.check_idp_availability",
-        return_value=None,
-    )
-
-    # set a default in case no marker is provided
-    init_view = None
-
-    marker = request.node.get_closest_marker("callback")
-    if marker and (_init_view := marker.kwargs.get("init_view")):
-        init_view = _init_view
-    if init_view is None:
-        raise TypeError("You must provide the init_view marker to the callback fixture")
-
-    response = init_view(auth_request)
-    redirect_url: str = response.url  # type: ignore
-    assert redirect_url
-    state_key = parse_qs(urlsplit(redirect_url).query)["state"][0]
-
-    callback_request = rf.get(
-        "/oidc/dummy-callback",
-        {"state": state_key, "code": "dummy-oidc-code"},
-    )
-    callback_request.session = auth_request.session
-    store_config(callback_request)
-
-    session = client.session
-    for key, value in callback_request.session.items():
-        session[key] = value
-    session.save()
-
-    return callback_request, client
 
 
 def test_oidc_enabled(settings):
